Add delete a blog

+++ part4/bloglist/backend/controllers/blogs.js
@@ -49,8 +49,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id){
+        return response.status(401).json({ error: 'Token Invalid' })
+    }
+
+    const blogToDelete = await Blog.findById(request.params.id)
+    const user = await User.findById(decodedToken.id)
+
+    if ( blogToDelete.user.toString() === user.id.toString()){
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
